docs: add 7guis example
